Issue 45483: allowed creation types in bulk insert modal to be based on if sample parent types exist and based on the creation type for the page
