+++ packages/components/src/entities/SampleCreatePage.tsx
@@ -1,16 +1,9 @@
-import React, { FC, memo, useCallback, useMemo, useState } from 'react';
+import React, { FC, memo, useCallback, useMemo } from 'react';
 import { WithRouterProps } from 'react-router';
-import { List, Map, OrderedMap } from 'immutable';
+import { OrderedMap } from 'immutable';
 import { Filter } from '@labkey/api';
 
-import {
-    ALIQUOT_CREATION,
-    CHILD_SAMPLE_CREATION,
-    DERIVATIVE_CREATION,
-    POOLED_SAMPLE_CREATION,
-    SampleCreationType,
-    SampleCreationTypeModel,
-} from '../internal/components/samples/models';
+import { SampleCreationType } from '../internal/components/samples/models';
 import { InjectedRouteLeaveProps, useRouteLeave } from '../internal/util/RouteLeave';
 import { useServerContext } from '../internal/components/base/ServerContext';
 import { useNotificationsContext } from '../internal/components/notifications/NotificationsContext';
@@ -22,7 +15,6 @@ import { SchemaQuery } from '../public/SchemaQuery';
 import { SCHEMAS } from '../internal/schemas';
 import { DataClassDataType, SampleTypeDataType } from '../internal/components/entities/constants';
 import { QueryInfo } from '../public/QueryInfo';
-import { EntityParentType } from '../internal/components/entities/models';
 import { BulkAddData } from '../internal/components/editable/EditableGrid';
 import { parseCsvString } from '../internal/util/utils';
 import { InsufficientPermissionsPage } from '../internal/components/permissions/InsufficientPermissionsPage';
@@ -39,31 +31,10 @@ import { MAX_EDITABLE_GRID_ROWS } from '../internal/constants';
 import { SampleTypeBasePage } from './SampleTypeBasePage';
 import { onSampleChange } from './actions';
 import { getSampleAuditBehaviorType, getSampleTypeTemplateUrl } from './utils';
-import { SAMPLES_LISTING_GRID_ID } from './SampleListingPage';
 import { useSampleTypeAppContext } from './SampleTypeAppContext';
 
 const TITLE = 'Sample Type';
 const SUBTITLE = 'Create New Samples';
-
-const getCreationTypes = (hasParentSamples: boolean, urlCreationType: string): SampleCreationTypeModel[] => {
-    if (!hasParentSamples) return [{ ...CHILD_SAMPLE_CREATION, selected: true }];
-
-    const types = [
-        {
-            ...DERIVATIVE_CREATION,
-            selected: !urlCreationType || urlCreationType === DERIVATIVE_CREATION.type,
-        },
-        { ...POOLED_SAMPLE_CREATION, selected: urlCreationType === POOLED_SAMPLE_CREATION.type },
-        { ...ALIQUOT_CREATION, selected: urlCreationType === ALIQUOT_CREATION.type },
-    ];
-
-    const selectedType = types.find(type => type.selected);
-    if (!selectedType) {
-        types[0] = { ...types[0], selected: true };
-    }
-
-    return types;
-};
 
 interface OwnProps {
     // loadNameExpressionOptions is a prop for testing purposes only
@@ -74,10 +45,7 @@ export interface SampleCreatePageProps extends OwnProps, CommonPageProps, WithRo
 
 export const SampleCreatePage: FC<SampleCreatePageProps> = memo(props => {
     const { loadNameExpressionOptions, location, navigate, router, routes, goBack } = props;
-    const { creationType, selectionKey, useAsync } = location?.query;
-    const [sampleCreationTypes, setSampleCreationTypes] = useState<SampleCreationTypeModel[]>(
-        selectionKey ? getCreationTypes(selectionKey.indexOf(SAMPLES_LISTING_GRID_ID) === 0, creationType) : []
-    );
+    const { useAsync } = location?.query;
     const { user } = useServerContext();
     const { createNotification } = useNotificationsContext();
     const { combineParentTypes, controllerName, downloadTemplateExcludeColumns, importHelpLinkTopic, parentDataTypes } =
@@ -151,16 +119,11 @@ export const SampleCreatePage: FC<SampleCreatePageProps> = memo(props => {
         [createNotification, navigate]
     );
 
-    const getFileTemplateUrl = useCallback((queryInfo: QueryInfo, importAliases: Record<string, string>): string => {
-        return getSampleTypeTemplateUrl(queryInfo, importAliases, downloadTemplateExcludeColumns);
-    }, []);
-
-    const onParentChange = useCallback(
-        (parentTypes: Map<string, List<EntityParentType>>): void => {
-            const numSampleParents = parentTypes.get(SCHEMAS.EXP_TABLES.SAMPLE_SETS.queryName)?.size;
-            setSampleCreationTypes(getCreationTypes(numSampleParents > 0, creationType));
+    const getFileTemplateUrl = useCallback(
+        (queryInfo: QueryInfo, importAliases: Record<string, string>): string => {
+            return getSampleTypeTemplateUrl(queryInfo, importAliases, downloadTemplateExcludeColumns);
         },
-        [creationType]
+        [downloadTemplateExcludeColumns]
     );
 
     const onBulkAdd = useCallback(
@@ -174,39 +137,37 @@ export const SampleCreatePage: FC<SampleCreatePageProps> = memo(props => {
             let pivotKey;
             let pivotValues = [];
             let haveMultiParent = false;
-            if (sampleCreationTypes?.length > 0) {
-                data.keySeq().forEach(key => {
-                    const isSampleParent = key.indexOf(SampleTypeDataType.insertColumnNamePrefix) === 0;
-                    const isDataClassParent = key.indexOf(DataClassDataType.insertColumnNamePrefix) === 0;
-                    if (isSampleParent || isDataClassParent) {
-                        if (data.get(key)) {
-                            const parents = data.get(key);
-                            if (parents.length > 0) {
-                                const values =
-                                    typeof parents[0] === 'string' ? parseCsvString(parents[0], ',') : parents;
-                                if (values.length > 1) {
-                                    if (haveMultiParent) {
-                                        validationMsg = combineParentTypes
-                                            ? 'Only one parent type with more than one value is allowed when creating non-pooled samples in bulk.'
-                                            : 'Only one source or parent with more than one value is allowed when creating non-pooled samples in bulk.';
-                                    } else if ((isSampleParent && !poolingSampleParents) || isDataClassParent) {
-                                        pivotValues = values;
-                                        pivotKey = key;
-                                        haveMultiParent = true;
-                                        totalItems = numItems * values.length;
-                                    }
+
+            data.keySeq().forEach(key => {
+                const isSampleParent = key.indexOf(SampleTypeDataType.insertColumnNamePrefix) === 0;
+                const isDataClassParent = key.indexOf(DataClassDataType.insertColumnNamePrefix) === 0;
+                if (isSampleParent || isDataClassParent) {
+                    if (data.get(key)) {
+                        const parents = data.get(key);
+                        if (parents.length > 0) {
+                            const values = typeof parents[0] === 'string' ? parseCsvString(parents[0], ',') : parents;
+                            if (values.length > 1) {
+                                if (haveMultiParent) {
+                                    validationMsg = combineParentTypes
+                                        ? 'Only one parent type with more than one value is allowed when creating non-pooled samples in bulk.'
+                                        : 'Only one source or parent with more than one value is allowed when creating non-pooled samples in bulk.';
+                                } else if ((isSampleParent && !poolingSampleParents) || isDataClassParent) {
+                                    pivotValues = values;
+                                    pivotKey = key;
+                                    haveMultiParent = true;
+                                    totalItems = numItems * values.length;
                                 }
                             }
                         }
                     }
-                });
-            }
+                }
+            });
 
             if (validationMsg) return { validationMsg };
             if (totalItems === 0) totalItems = numItems;
             return { pivotKey, pivotValues, totalItems };
         },
-        [sampleCreationTypes?.length]
+        [combineParentTypes]
     );
 
     if (!user.hasInsertPermission()) {
@@ -225,7 +186,6 @@ export const SampleCreatePage: FC<SampleCreatePageProps> = memo(props => {
                 asyncSize={useAsync === 'true' ? 1 : BACKGROUND_IMPORT_MIN_FILE_SIZE}
                 auditBehavior={auditBehavior}
                 canEditEntityTypeDetails={user.hasDesignSampleTypesPermission()}
-                creationTypeOptions={sampleCreationTypes}
                 combineParentTypes={combineParentTypes}
                 containerFilter={getContainerFilterForLookups()}
                 entityDataType={entityDataType}
@@ -243,7 +203,6 @@ export const SampleCreatePage: FC<SampleCreatePageProps> = memo(props => {
                 onBackgroundJobStart={onBackgroundJobStart}
                 onBulkAdd={onBulkAdd}
                 onCancel={goBack}
-                onParentChange={onParentChange}
                 parentDataTypes={parentDataTypes}
                 setIsDirty={setIsDirty}
             />
