+++ packages/components/src/entities/index.ts
@@ -68,12 +68,10 @@ import {
     SampleListingPage,
     getSamplesImportSuccessMessage,
     getSamplesCreatedSuccessMessage,
-    SAMPLES_LISTING_GRID_ID,
 } from './SampleListingPage';
 import { SampleCreatePage } from './SampleCreatePage';
 
 export {
-    SAMPLES_LISTING_GRID_ID,
     createEntityParentKey,
     downloadSampleTypeTemplate,
     filterMediaSampleTypes,
